+++ src/RendererFactory.ts
@@ -7,23 +7,26 @@ import Schematic from './Schematic';
 import Builder from './Builder';
 import Renderer from './Renderer';
 import FlowRenderer from './renderers/Flow';
-import ReactRenderer from './renderers/React';
+import PropTypesRenderer from './renderers/PropTypes';
 import TypeScriptRenderer from './renderers/TypeScript';
-import { Options } from './types';
+import { Options, RendererType } from './types';
 
 export default class RendererFactory {
   /**
    * Create a new renderer with the defined schematic.
    */
-  static factory(options: Options, builder: Builder, schematic: Schematic): Renderer {
-    const { renderer } = options;
-
+  static factory(
+    renderer: RendererType,
+    options: Options,
+    builder: Builder,
+    schematic: Schematic,
+  ): Renderer {
     switch (renderer.toLowerCase()) {
-      case 'react':
-        return new ReactRenderer(options, builder, schematic);
-
       case 'flow':
         return new FlowRenderer(options, builder, schematic);
+
+      case 'prop-types':
+        return new PropTypesRenderer(options, builder, schematic);
 
       case 'typescript':
         return new TypeScriptRenderer(options, builder, schematic);