+++ tests/RendererFactory.test.js
@@ -1,5 +1,5 @@
 import FlowRenderer from '../src/renderers/Flow';
-import ReactRenderer from '../src/renderers/React';
+import PropTypesRenderer from '../src/renderers/PropTypes';
 import RendererFactory from '../src/RendererFactory';
 import TypeScriptRenderer from '../src/renderers/TypeScript';
 import { options } from './mocks';
@@ -7,35 +7,17 @@ import { options } from './mocks';
 describe('RendererFactory', () => {
   describe('factory()', () => {
     it('factories renderer objects', () => {
-      expect(
-        RendererFactory.factory({
-          ...options,
-          renderer: 'flow',
-        }),
-      ).toBeInstanceOf(FlowRenderer);
+      expect(RendererFactory.factory('flow', options)).toBeInstanceOf(FlowRenderer);
 
-      expect(
-        RendererFactory.factory({
-          ...options,
-          renderer: 'ReAcT',
-        }),
-      ).toBeInstanceOf(ReactRenderer);
+      expect(RendererFactory.factory('prop-TYPES', options)).toBeInstanceOf(PropTypesRenderer);
 
-      expect(
-        RendererFactory.factory({
-          ...options,
-          renderer: 'TYPESCRIPT',
-        }),
-      ).toBeInstanceOf(TypeScriptRenderer);
+      expect(RendererFactory.factory('typeSCRIPT', options)).toBeInstanceOf(TypeScriptRenderer);
     });
 
     it('errors for invalid renderer', () => {
-      expect(() =>
-        RendererFactory.factory({
-          ...options,
-          renderer: 'foo',
-        }),
-      ).toThrowError('Renderer "foo" not supported.');
+      expect(() => RendererFactory.factory('foo', options)).toThrowError(
+        'Renderer "foo" not supported.',
+      );
     });
   });
 });